Add the Big-Ass Debugger for Lightrec
The Big-Ass Debugger will run two instances of PCSX in parallel.
The first one will run an instance of PCSX running with Lightrec, the
second one with the interpreter. Each instance will provide the debugger
with checksums of the registers and memory after each trace is executed.

This allows to detect any misbehaving recompiled code: it permits to see
the original PSX code, see the recompiled version, as well as the
differences in the emulator state that triggered the debugger.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>