lightrec: Allocate and provide a code buffer to Lightrec
Provide a 8 MiB area in which Lightrec will be able to store the
generated code. Map it to (offset + 0x10000000) by default.

With the new memory init functions, the start and end addresses of this
code buffer are 32-bit addresses, which mean that all pointers to
generated functions will have 32-bit addresses as well. Lightrec can
then shrink the code LUT from 5 MiB to 2.5 MiB on 64-bit platforms.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>